requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-treebeard==4.5.1
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sentry-sdk==1.1.0
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.5
djangorestframework~=3.12.4